Alhambra::LION_FOUNTAIN -> Alhambra.lion_fountain
Load order was causing problems with the Tile model not properly loading
before setting Alhambra::LION_FOUNTAIN, but yet not raising an error
about an unknown constant. In short, the constant (LION_FOUNTAIN) was
being set to nil. Requiring 'tile' at the top of the module file didn't
help. So we'll just make it a run time method instead of a load time
constant.